+++ app/Http/Controllers/StudentController.php
@@ -288,10 +288,20 @@ class StudentController extends Controller
 
         $student->unsetRelation('grades');
         $student->grades = $student->grades()->whereNull('from_student_id')->get();
-
+        $student->load(['logEntries' => function($query) {
+            $query->orderByDesc('id')->take(100);
+        }]);
         return view('students.show', compact('student', 'section', 'shop', 'class', 'admin', 'items', 'challenges', 'cards', 'evaluation', 'settings', 'allcards'));
     }
 
+    public function getStudentLog($id)
+    {
+        $student = Student::find($id);
+        $class = Classroom::find($student->classroom->classroom_id);
+        $this->authorize('studyOrTeach', $class);
+        return $student->logEntries;
+
+    }
     public static function getIndividualReport($id, $class)
     {
         $student = Student::findOrFail($id);